Rearrangement of code as a possible common basis.
Used different commits of branch `proto-bude-fw-spawn-cars` as source
for a new game module. the demo module will be deleted in further
commits.

Reworked physics system using `ExternalForce`, `Friction`, and an own
air friction. Now `Avian` handles the speed of the cars with is what it was designed for. Next steps: Movement using an external force, maybe with
an initial velocity; Explosions through ExternalImpulse.

Made a ground plane with a low friction. It has a non-moving car for
testing purposes. Next steps: Rotate the used coordinate system 100°
around the Y-Axis to have a more convinient handling of directions (x is
right, z into the camera).