Added the creation of car colliders.
They are not used nor tested yet. The car body and the 4 wheels have one
collider each. Every wheel has a `Transform`, too. This is necessary,
because the colliders for the wheels are untranslatable cyclinders
and the anchor point of the entity is not the center of the mesh.
To use the wheel colliders, one has to spawn them as child entities to the
root entity (with the body collider) and give them useful tags. Do not
spwan them on the scene when it is ready, because then the car could
exist for a few frames without physics. The transforms of those child
entities are the offsets from the corresponding mesh center to the entity anchor.